Lexicon lookup return a deepcopy of the lexicon word
The reason behind this behaviour is that each word is independant
from other occurences of the same word. For example, the word "si"
would be realised as "s'" if it was followed by (for example) the
word "ils", but not if it was followed by the word "elles".

Changing the realisation of a word should not impact other occurences
of the same word elsewhere.